src/requirements.txt
python-dotenv
aiohttp
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
python-pptx
openai>=1.0.0